MAININTRO: split done for part 3
