+++ src/MAININTRO.c
@@ -9,7 +9,7 @@ uint32  IntroMemL = 0;
 
 // ------------------------------
 
-void SETDARKCOLOR1(char* FName, r_Col_t* Colors)
+void SETDARKCOLOR(char* FName, r_Col_t* Colors)
 {
     uint32  AddrX, AddrEnd, ISize;
     uint8   i;
@@ -136,8 +136,6 @@ void MAININTRO()
     uint32      SMemL[3];
     char        s[40];
     char*       _s;
-    BPTR        FHandle;
-    r_Col_t     Colors[128];
     PLANEPTR    MyRastPtr = NULL;
     struct TmpRas       MyTmpRas;
     struct AreaInfo     MyAI;
@@ -164,7 +162,6 @@ void MAININTRO()
     double      FacSin, FacCos;
     double      SizeFactor;
     double      Factor;
-    uint16      dFactor;
     double      store;
     uint8       i, k;
     sint16      SXdata0int, SYdata0int;
@@ -185,170 +182,14 @@ void MAININTRO()
     if      (0 > part_return) { goto leave_intro; }
     else if (0 < part_return) { goto leave_intro; }
 
-
     part_return = MAININTRO_PART2(SMemA, SMemL);
     if      (0 > part_return) { goto leave_intro; }
     else if (0 < part_return) { goto leave_intro; }
 
+    part_return = MAININTRO_PART3(SMemA, SMemL);
+    if      (0 > part_return) { goto leave_intro; }
+    else if (0 < part_return) { goto leave_intro; }
 
-    /*****************************************************************************/
-    // prepare next effect...
-    IMemL[0] = 201856;
-    IMemL[1] = 21000;
-    for (i = 0; i<2; ++i)
-    {
-        IMemA[i] = (uint8*) AllocMem(IMemL[i], MEMF_CHIP);
-        if (NULL == IMemA[i])
-        {
-            goto leave_intro;
-        }
-    }
-
-    SPAddrC = SMemA[1]+(SMemL[1]/2)+6500; SPVolC = 64; SPFreqC = 380; SPLengthC = (SMemL[1]-13000)/4;
-    SPAddrD = SMemA[1]+6500;              SPVolD = 64; SPFreqD = 380; SPLengthD = (SMemL[1]-13000)/4;
-
-    custom.dmacon = BITSET | DMAF_AUD2 | DMAF_AUD3; // 0x800C
-    WaitTOF();
-    WaitTOF();
-
-    SPLengthD = 1;
-    SPLengthC = 1;
-
-    i = 255;
-    do
-    {
-        SetRGB32(MyVPort_PTR[1], 0, i<<24, i<<24, i<<24);
-        i = i-5;
-        WaitTOF();
-    }
-    while (i > 10);
-    SetRGB32(MyVPort_PTR[1],0,0,0,0);
-    custom.dmacon = BITCLR | DMAF_AUD2 | DMAF_AUD3; // 0x000C
-    if (LMB_PRESSED)
-        { goto leave_intro; }
-
-    (void) my_strcpy(_s, "MOD.Intro");
-    FHandle = OPENSMOOTH(s,MODE_OLDFILE);
-    if (0 != FHandle)
-    {
-        Close(FHandle);
-        StopPlayer();
-        FreePlayer();
-        SndModulePtr = LoadModule(s);   // module will be loaded directly from disk
-        if (GETMIDIPLAYER(SndModulePtr))
-        {
-            PlayModule(SndModulePtr);
-        }
-    }
-
-    (void) my_strcpy(_s, "Frame4.pal");
-    SETDARKCOLOR1(s, Colors);
-    (void) my_strcpy(_s, "Frame4.img");       // terrain with stars...
-    if (!DISPLAYIMAGE(s,0,75,640,360,7,MyScreen[1],0))
-    {
-        goto leave_intro;
-    }
-    ClipBlit(MyRPort_PTR[1],0,75,MyRPort_PTR[0],0,75,640,360,192);
-
-    (void) my_strcpy(_s, "Frame5.img");       // little earth-image..
-
-    FHandle = OPENSMOOTH(s,MODE_OLDFILE);
-    if (0 != FHandle)
-    {
-        (void)  Seek(FHandle, 0, OFFSET_END);
-        ISize = Seek(FHandle, 0, OFFSET_BEGINNING);
-        (void) Read(FHandle,(APTR) (IMemA[0]+IMemL[0]-ISize-250), ISize);
-        UNPACK(IMemA[0],IMemA[0]+IMemL[0]-ISize-250,20160,0);
-        Close(FHandle);
-    }
-
-    struct BitMap MyBitMap = { 20, 144, 1, 7, 0, \
-                            {(PLANEPTR) (IMemA[0]),      (PLANEPTR) (IMemA[0]+2880), (PLANEPTR) (IMemA[0]+5760), \
-                             (PLANEPTR) (IMemA[0]+8640), (PLANEPTR) (IMemA[0]+11520),(PLANEPTR) (IMemA[0]+14400), \
-                             (PLANEPTR) (IMemA[0]+17280), NULL }};
-
-    // fade in....
-    dFactor = 0;
-    k = 20;
-    do
-    {
-        ScreenToFront(MyScreen[AScr]);
-        AScr=1-AScr;
-        dFactor += 0xCD; // 11001101 =~ 0,05  >>12
-        for (i = 1; i<128; ++i)
-        {
-            SetRGB32(MyVPort_PTR[AScr], i, (Colors[i].r*dFactor)<<12,
-                                           (Colors[i].g*dFactor)<<12,
-                                           (Colors[i].b*dFactor)<<12);
-        }
-        WaitTOF();
-        --k;
-    }
-    while (0 != k);
-
-    // set all colours to the max. values
-    for (i = 1; i<128; ++i)
-    {
-        SetRGB32(MyVPort_PTR[1-AScr], i, Colors[i].r<<24, Colors[i].g<<24, Colors[i].b<<24);
-        SetRGB32(MyVPort_PTR[  AScr], i, Colors[i].r<<24, Colors[i].g<<24, Colors[i].b<<24);
-    }
-    ScrollRaster(MyRPort_PTR[AScr],0,-4,0,75,639,434);
-    WaitTOF();
-
-    for (i = 0; i<6; ++i)
-    {
-        ScreenToFront(MyScreen[AScr]);
-        AScr = 1-AScr;
-        actRastPort = MyRPort_PTR[AScr];
-        ScrollRaster(actRastPort,0,-8,0,75,639,434);
-        WaitTOF();
-        ClipBlit(actRastPort,0,270,actRastPort,0,75,640,8,192);
-        if (LMB_PRESSED)
-            { goto leave_intro; }
-    }
-
-    for (i = 0; i<35; ++i)
-    {
-        ScreenToFront(MyScreen[AScr]);
-        AScr = 1-AScr;
-        actRastPort = MyRPort_PTR[AScr];
-        ScrollRaster(actRastPort,0,-8,0,75,639,434);
-        ClipBlit(actRastPort,0,270,actRastPort,0,75,640,8,192);
-        BltBitMapRastPort(&MyBitMap,0,136-i*4,actRastPort,380,75,160,8,192);
-        WaitTOF();
-        if (LMB_PRESSED)
-            { goto leave_intro; }
-    }
-
-    ScreenToFront(MyScreen[AScr]);
-    AScr = 1-AScr;
-    actRastPort = MyRPort_PTR[AScr];
-    ScrollRaster(actRastPort,0,-8,0,75,639,434);
-    ClipBlit(actRastPort,0,270,actRastPort,0,75,640,8,192);
-    BltBitMapRastPort(&MyBitMap,0,0,actRastPort,380,79,160,4,192);
-    WaitTOF();
-
-    for (i = 0; i<14; ++i)
-    {
-        ScreenToFront(MyScreen[AScr]);
-        AScr = 1-AScr;
-        actRastPort = MyRPort_PTR[AScr];
-        ScrollRaster(actRastPort,0,-8,0,75,639,434);
-        WaitTOF();
-        ClipBlit(actRastPort,0,270,actRastPort,0,75,640,8,192);
-        if (LMB_PRESSED)
-            { goto leave_intro; }
-    }
-
-    Delay(10);
-    if (LMB_PRESSED)
-        { goto leave_intro; }
-
-    SetRast(actRastPort, 0); // clear Rastport
-    ScreenToFront(MyScreen[AScr]);
-    AScr = 1-AScr;
-    actRastPort = MyRPort_PTR[AScr];
-    SetRast(actRastPort, 0); // clear Rastport
 
     (void) my_strcpy(_s, "Frame6.pal");
     (void) SETCOLOR(MyScreen[  AScr],s);
@@ -362,7 +203,7 @@ void MAININTRO()
     (void) my_strcpy(_s, "Frame6.img");
     if (!DISPLAYIMAGE(s,0,75,640,360,7,MyScreen[AScr],0)) { goto leave_intro; }
     ScreenToFront(MyScreen[AScr]);
-    MyBitMap = (struct BitMap) { 80, 360, 1, 7, 0, \
+    struct BitMap MyBitMap = (struct BitMap) { 80, 360, 1, 7, 0, \
                             {(PLANEPTR) (IMemA[0]),       (PLANEPTR) (IMemA[0]+28800), (PLANEPTR) (IMemA[0]+57600), \
                              (PLANEPTR) (IMemA[0]+86400), (PLANEPTR) (IMemA[0]+115200),(PLANEPTR) (IMemA[0]+144000), \
                              (PLANEPTR) (IMemA[0]+172800)}};
